Team is created & deleted based on the player events PlayerLoggedInEvent and PlayerLoggedOutEvent, #616.

+++ src/main/java/bassebombecraft/event/entity/team/TeamMembershipEventHandler.java
@@ -7,6 +7,8 @@ import static bassebombecraft.player.PlayerUtils.isTypePlayerEntity;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraftforge.event.entity.living.LivingDeathEvent;
+import net.minecraftforge.event.entity.player.PlayerEvent.PlayerLoggedInEvent;
+import net.minecraftforge.event.entity.player.PlayerEvent.PlayerLoggedOutEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 /**
@@ -17,8 +19,6 @@ public class TeamMembershipEventHandler {
 
 	@SubscribeEvent
 	static public void handleLivingDeathEvent(LivingDeathEvent event) {
-		
-		// get repository
 		TeamRepository repository = getBassebombeCraft().getTeamRepository();
 
 		// remove living entity from team upon death
@@ -31,9 +31,31 @@ public class TeamMembershipEventHandler {
 		// disband team if dead entity is commander
 		if (isTypePlayerEntity(event.getEntityLiving())) {
 			PlayerEntity player = (PlayerEntity) event.getEntityLiving();			
+			
+			// exit if player isn't a commander
+			if(!repository.isCommander(player)) return;
+			
+			// disband team
 			repository.deleteTeam(player);
 			return;
 		}
 		
 	}
+
+	@SubscribeEvent
+	static public void handlePlayerLoggedInEvent(PlayerLoggedInEvent event) {
+
+		// register team
+		TeamRepository repository = getBassebombeCraft().getTeamRepository();		
+		repository.createTeam(event.getPlayer());
+	}
+	
+	@SubscribeEvent
+	static public void handlePlayerLoggedOutEvent(PlayerLoggedOutEvent event) {
+
+		// disband team
+		TeamRepository repository = getBassebombeCraft().getTeamRepository();		
+		repository.deleteTeam(event.getPlayer());
+	}
+		
 }
